[v10r2] Python 3 client fixes

See merge request lhcb-dirac/LHCbDIRAC!890

+++ LHCbDIRAC/Core/Utilities/ProductionData.py
@@ -15,6 +15,9 @@ The methods here are mostly from ancient history and need to be
 reviewed, these methods were grouped together as they form the
 "interface" for production clients and workflow modules to create LFNs.
 """
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
 
 __RCSID__ = "$Id$"
 
@@ -228,7 +231,7 @@ def constructUserLFNs(jobID, owner, outputFiles, outputPath='', prependString=''
     for outputFile in outputFiles:
       outputFile = outputFile.replace('LFN:', '')
       lfn = os.sep + os.path.join('lhcb', 'user', initial, owner, outputPath, yearMonth,
-                                  str(jobID / 1000), str(jobID)) + os.sep + os.path.basename(outputFile)
+                                  str(int(jobID / 1000)), str(jobID)) + os.sep + os.path.basename(outputFile)
       outputLFNs[outputFile] = lfn
 
   outputData = outputLFNs.values()
@@ -277,7 +280,7 @@ def _makeProductionPath(jobID, lfnROOT, typeName, prodstring, log=False):
   if log:
     try:
       jobid = int(jobID)
-      jobindex = str(jobid / 10000).zfill(4)
+      jobindex = str(int(jobid / 10000)).zfill(4)
     except Exception:
       jobindex = '0000'
     result += jobindex
@@ -297,7 +300,7 @@ def _makeProductionLFN(jobID, lfnROOT, filetuple, prodstring):
                                                                                             str(filetuple)))
   try:
     jobid = int(jobID)
-    jobindex = str(jobid / 10000).zfill(4)
+    jobindex = str(int(jobid / 10000)).zfill(4)
   except Exception:
     jobindex = '0000'
