+++ tests/System/GridTestSubmission/exe-script.py
@@ -10,6 +10,9 @@
 # or submit itself to any jurisdiction.                                       #
 ###############################################################################
 '''Script to run Executable application'''
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
 
 from os import system
 import sys
@@ -18,4 +21,8 @@ import string
 # Main
 if __name__ == '__main__':
 
-  sys.exit(system('''echo Hello World %s''' % (''.join(random.choice(string.ascii_letters) for _ in range(10)))) / 256)
+  sys.exit(int(system(
+      '''echo Hello World %s''' % (
+          ''.join(random.choice(string.ascii_letters) for _ in range(10))
+      )
+  ) / 256))
