+++ tests/Workflow/Integration/gaudi-script.py
@@ -10,6 +10,9 @@
 # or submit itself to any jurisdiction.                                       #
 ###############################################################################
 '''Script to run '''
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
 
 from os import system, environ, pathsep, getcwd
 import sys
@@ -27,4 +30,4 @@ if __name__ == '__main__':
 
   options = optGauss + optDec + optPythia + optOpts + optCompr + optPConf
   LHCbApp().EvtMax = 2
-  sys.exit(system('''gaudirun.py -T %s''' % options) / 256)
+  sys.exit(int(system('''gaudirun.py -T %s''' % options) / 256))
